+++ packages/bds_colors/lib/bds_colors.dart
@@ -1,13 +1,16 @@
 /// Flutter package for BDS Colors.
 library bds_colors;
 
-export 'package:bds_colors/palettes/base_colors.g.dart';
-export 'package:bds_colors/palettes/palettes.g.dart';
-export 'package:bds_colors/themes/themes.dart';
+export 'package:bds_colors/data/base_colors.g.dart';
+export 'package:bds_colors/data/palettes.g.dart';
+export 'package:bds_colors/data/themes.dart';
+
+export 'package:bds_colors/palette/color_palette.dart';
+export 'package:bds_colors/palette/shade.dart';
+
+export 'package:bds_colors/theme/color_theme.dart';
+export 'package:bds_colors/theme/token.dart';
 
 export 'package:bds_colors/widgets/colors_data.dart';
-
 export 'package:bds_colors/color_context.dart';
-export 'package:bds_colors/color_palette.dart';
-export 'package:bds_colors/color_theme.dart';
 export 'package:bds_colors/smart_color.dart';
