+++ packages/bds_colors/lib/theme/token.dart
@@ -1,0 +1,11 @@
+/// Tokens for [ColorTheme].
+enum Token {
+  background,
+  backgroundHover,
+  backgroundActive,
+  backgroundSelected,
+  backgroundSelectedHover,
+  backgroundBrand,
+  backgroundInverse,
+  backgroundInverseHover,
+}
