+++ packages/bds_colors/lib/widgets/colors_data.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/widgets.dart';
 
-import 'package:bds_colors/themes/themes.dart';
-import 'package:bds_colors/widgets/palette_scope.dart';
-import 'package:bds_colors/widgets/theme_scope.dart';
+import 'package:bds_colors/data/themes.dart';
 
-import 'package:bds_colors/color_theme.dart';
+import 'package:bds_colors/palette/palette_scope.dart';
+import 'package:bds_colors/theme/color_theme.dart';
+import 'package:bds_colors/theme/theme_scope.dart';
 import 'package:bds_colors/smart_color.dart';
 
 class ColorsData extends StatelessWidget {
